feat(RHINENG-13864): Load activation keys
This PR adds the ability to load activation keys from the Activation Keys UI app and allows a user to select one of those activation keys to get a tailored registration assistant experience.

When an activation key is selected, a dropdown should appear below the select box giving details on the key. And stepping through the Progress Stepper should eventually show the activation key name in the Code Block for the commands

A user may also auto-generate a key from the dropdown menu. This will create a key with the name `activation-key-<timestamp>` which will be automatically selected for the user.