fix(RHINENG-18539): Activation keys empty state
We had a bad conditional for showing when a user has no activation keys.
To test, go to Registration Assistant using an account with no
activation keys (I can provide a login if you don't have one). Click the
activation keys selection dropdown and wait for the fetch to complete.